app/trip_generator/requirements.txt
fastapi
uvicorn
python-docx
docx2pdf
openai
pydantic
docxtpl
jinja2
